Refactor

+++ src/components/Tag/index.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useEffect, useState } from 'react'
+import { FC, useEffect, useState } from 'react'
 import { ITag } from '../../../interfaces'
 import { faCheck, faXmark } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
@@ -10,7 +10,7 @@ type PropsType = {
 	isEdit?: boolean
 	isSelected?: boolean
 	onToggle?: (tag: ITag, isChecked: boolean) => void
-	onRemove?: (tag: ITag, isChecked: boolean) => void
+	onRemove?: (tag: ITag) => void
 }
 
 const Tag: FC<PropsType> = ({
@@ -27,7 +27,7 @@ const Tag: FC<PropsType> = ({
 			onToggle(tag, isChecked)
 			setIsChecked(!isChecked)
 		} else {
-			onRemove(tag)
+			onRemove && onRemove(tag)
 		}
 	}
